src/openssl_engine/requirements.txt
pyopenssl
cryptography>=36